requirements.txt
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
paho-mqtt==1.6.1
numpy==1.24.3
